pyproject.toml
[tool.poetry]
name = "ref-drf"
version = "0.1.0"
description = ""
authors = ["grbcas <grbcas@gmail.com>"]
readme = "README.md"
packages = [{include = "ref_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-debug-toolbar = "^4.2.0"
users = "^1.0.dev0"
celery = "^5.3.6"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
